Move the location of the reivew shorten service
