+++ packages/review/src/components/shorten-list/popular-reviews.tsx
@@ -1,7 +1,7 @@
-import { usePopularReviews } from '../../services'
 import type { SortingType } from '../sorting-context'
 
 import { ReviewsList } from './reviews-list'
+import { usePopularReviews } from './services'
 import type { ShortenReview } from './types'
 
 export function PopularReviews({
